requirements.txt
numpy==1.24.3
opencv-python-headless==4.8.1.78
Pillow==10.0.1
scikit-learn==1.3.2
joblib==1.3.2
boto3==1.34.0
botocore==1.34.0
requests==2.32.4